set rostersize back to separate selector
